[UNSTABLE][NEEDS TESTING] Flatten callback pyramid in devices/actions.ts
