+++ webpack/devices/devices.tsx
@@ -12,6 +12,7 @@ import {
 import { Diagnosis, DiagnosisName } from "./connectivity/diagnosis";
 import { StatusRowProps } from "./connectivity/connectivity_row";
 import { refresh } from "../api/crud";
+import { devices } from "../device";
 
 @connect(mapStateToProps)
 export class Devices extends React.Component<Props, {}> {
@@ -43,9 +44,8 @@ export class Devices extends React.Component<Props, {}> {
   }
 
   refresh = () => {
-    this
-      .props
-      .dispatch(refresh(this.props.deviceAccount));
+    this.props.dispatch(refresh(this.props.deviceAccount));
+    devices.current.getState();
   };
 
   render() {
